Random

Documentation

data RandomResult a g where

The result object to pass around between functions that depend on random state

Constructors

RandResult :: RandomGen g => (a, g) -> RandomResult a g

randomFn :: (Random a, Ord a, RandomGen g) => g -> (a, a) -> RandomResult a g

The random function that produces a random value between given low and high values. It reorders the given range.

rfmap :: RandomGen g => (a -> b) -> RandomResult a g -> RandomResult b g

Equivalent of fmap but for random results. It simply applies the given function to the result without changing its random state.

randomChain :: (Ord a, RandomGen g) => RandomResult a g -> (g -> (a, a) -> RandomResult a g) -> (a, a) -> RandomResult a g

given a random result and a function that makes random result from two-tuple and a two-tuple make a random result

liftRandVal :: RandomGen g => RandomResult a g -> a

liftRandGen :: RandomGen g => RandomResult a g -> g

randFoldl :: RandomGen g => g -> NonEmptyList (g -> (a, a) -> RandomResult a g, (a, a)) -> RandomResult (NonEmptyList a) g

randFoldlFixedRange :: RandomGen g => g -> (a, a) -> NonEmptyList (g -> (a, a) -> RandomResult a g) -> RandomResult (NonEmptyList a) g

randFoldlFixedRange2 :: RandomGen g => g -> NonEmptyList (g -> RandomResult a g) -> RandomResult (NonEmptyList a) g

randMap :: (Ord a, RandomGen g) => g -> (g -> (a, a) -> RandomResult a g) -> NonEmptyList (a, a) -> RandomResult (NonEmptyList a) g

randomDouble :: RandomGen g => g -> (Double, Double) -> RandomResult Double g

randomInt :: RandomGen g => g -> (Int, Int) -> RandomResult Int g

randomWord :: RandomGen g => g -> (Word, Word) -> RandomResult Word g

randomWaveVal :: RandomGen g => g -> (WaveVal, WaveVal) -> RandomResult WaveVal g

randomPowerVal :: RandomGen g => g -> (PowerVal, PowerVal) -> RandomResult PowerVal g

randval :: RandomGen g => g -> RandomResult Double g